item.h: make weight a signed long
This avoids having to do overflow checks when reading from the
configuration file.

+++ item.h
@@ -5,7 +5,7 @@
 
 struct item {
 	char *name;
-	int weight;
+	long weight;
 	struct list_head list;
 };
